Create separate bundle for attribute-editor
