+++ cvat/apps/engine/templates/engine/annotation.html
@@ -486,5 +486,5 @@
 {% endblock %}
 
 {% block footer %}
-    <script src="/cvat-ui.min.js" type="text/javascript"></script>
+    <script src="/attribute-editor.min.js" type="text/javascript"></script>
 {% endblock %}
